pyproject.toml
[project]
name = "music-lyrics-tagger"
version = "0.1"
description = ""
readme = "README.md"
authors = [{name = "Karn Wong"}]
requires-python = ">=3.13"
dependencies = [
    "azapi>=3.0.8",
    "click",
    "lyricsgenius>=3.2.0",
    "mutagen>=1.47.0",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pytest>=8.3.4",
]

[project.urls]
Homepage = "https://github.com/kahnwong/music-lyrics-tagger"

[project.scripts]
music-lyrics-tagger = "music_lyrics_tagger.cli:cli"

[project.optional-dependencies]
test = ["pytest"]
